# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Range Monitor API",
    "summary": "The backend for the Range Monitor",
...
    "version": "0.1.3"
  },
  "paths": {
    "/auth/login/": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login User",
        "description": "Authenticates a user and issues a JWT token pair upon successful login.\nThe access token is to be used for authenticating API requests, while the refresh\ntoken can be used to obtain new access tokens when the current one expires.",
        "operationId": "login_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid username or password",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh/": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Refresh Authentication",
        "description": "**protected**\nRefreshes an access token using a valid refresh token\nreturning the rotated token claims.",
        "operationId": "refresh_authentication",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenClaim"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/token": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get Token Details",
        "description": "**protected**\nRetrieves details about the currently authenticated access token.\nThe provided token must be a valid access token.",
        "operationId": "get_token_details",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2Token": []
          }
        ]
      }
    },
    "/auth/logout/": {
      "delete": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logout User",
        "description": "**protected**\nRevokes the given refresh token and its associated access token.\nThis effectively logs the user out by invalidating their current\nsession tokens.",
        "operationId": "logout_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation error",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Invalid or expired session",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/{user_id}/": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Read User",
        "description": "**Admin Role Required**\nRetrieves a user by their unique ID.",
        "operationId": "read_user",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "The unique identifier of the user.",
              "title": "User Id"
            },
            "description": "The unique identifier of the user."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "User is not an admin.",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User ID provided does not exist.",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete User",
        "description": "**Admin Role Required**\nDeletes a user and lazy deletes the token claims\nby incrementing the `credential_version`.",
        "operationId": "delete_user",
        "security": [
          {
            "OAuth2Token": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "The unique identifier of the user.",
              "title": "User Id"
            },
            "description": "The unique identifier of the user."
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation error",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "User is not an admin or tries delete self",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User ID provided does not exist.",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Patch User",
        "description": "**Admin Role Required**\nUpdates a user's information. If the users\nRole or password changes, the users `credential_version` is\nincremented.",
        "operationId": "patch_user",
        "security": [
          {
            "OAuth2Token": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "The unique identifier of the user.",
              "title": "User Id"
            },
            "description": "The unique identifier of the user."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPatchBody"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "User is not an admin.",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User ID provided does not exist.",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "errorType": {
            "type": "string",
            "title": "Errortype",
            "description": "A string representing the type of error that occurred",
            "default": "about:blank"
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "description": "The HTTP status code of the error response"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string",
                "description": "A detailed description of the error"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "requestId": {
            "anyOf": [
              {
                "type": "string",
                "description": "A Correlation ID of the request."
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestid"
          },
          "code": {
            "anyOf": [
              {
                "type": "string",
                "description": "An application-specific error code"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "extras": {
            "anyOf": [
              {
                "type": "object",
                "description": "A dictionary of additional error details"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extras"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ErrorResponse",
        "description": "The response schema for errors"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 128,
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9_.-]+$",
            "title": "Username",
            "description": "The username of the user."
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 3,
            "title": "Password",
            "description": "The password of the user."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "RefreshRequest": {
        "properties": {
          "refreshToken": {
            "type": "string",
            "title": "Refreshtoken",
            "description": "The refresh token used to obtain a new access token."
          },
          "accessToken": {
            "type": "string",
            "title": "Accesstoken",
            "description": "The current access token, if available. This is used to verify the session and ensure the refresh token is valid."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "refreshToken",
          "accessToken"
        ],
        "title": "RefreshRequest"
      },
...
      "TokenClaim": {
        "properties": {
          "accessToken": {
            "type": "string",
            "title": "Accesstoken"
          },
          "refreshToken": {
            "type": "string",
            "title": "Refreshtoken"
          },
          "tokenType": {
            "type": "string",
            "const": "bearer",
            "title": "Tokentype",
            "default": "bearer"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "title": "Expiresat"
          },
          "issuedAt": {
            "type": "integer",
            "title": "Issuedat"
          }
        },
        "type": "object",
        "required": [
          "accessToken",
          "refreshToken",
          "expiresAt",
          "issuedAt"
        ],
        "title": "TokenClaim"
      },
      "TokenDetails": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/TokenUser"
          },
          "issuedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Issuedat"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "title": "Expiresat"
          },
          "timeToLive": {
            "type": "integer",
            "title": "Timetolive"
          }
        },
        "type": "object",
        "required": [
          "user",
          "issuedAt",
          "expiresAt",
          "timeToLive"
        ],
        "title": "TokenDetails"
      },
      "TokenRequestBody": {
        "properties": {
          "refreshToken": {
            "type": "string",
            "title": "Refreshtoken",
            "description": "The refresh token used to obtain a new access token."
          },
          "accessToken": {
            "type": "string",
            "title": "Accesstoken",
            "description": "The current access token, if available. This is used to verify the session and ensure the refresh token is valid."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "refreshToken",
          "accessToken"
        ],
        "title": "TokenRequestBody"
      },
      "TokenResponse": {
        "properties": {
          "claim": {
            "$ref": "#/components/schemas/TokenClaim"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "title": "Userid",
            "description": "The unique identifier of the user."
          },
          "username": {
            "type": "string",
            "maxLength": 128,
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9_.-]+$",
            "title": "Username",
            "description": "The username of the user."
          },
          "role": {
            "$ref": "#/components/schemas/UserRoles",
            "description": "The role of the user."
          }
        },
        "type": "object",
        "required": [
          "claim",
          "userId",
          "username",
          "role"
        ],
        "title": "TokenResponse"
      },
      "TokenUser": {
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "title": "Userid",
            "description": "The unique identifier of the user."
          },
          "username": {
            "type": "string",
            "maxLength": 128,
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9_.-]+$",
            "title": "Username",
            "description": "The username of the user."
          },
          "cver": {
            "type": "integer",
            "title": "Cver"
          },
          "role": {
            "$ref": "#/components/schemas/UserRoles",
            "description": "The role of the user."
          }
        },
        "type": "object",
        "required": [
          "userId",
          "username",
          "cver",
          "role"
        ],
        "title": "TokenUser"
      },
...
      "UserPatchBody": {
        "properties": {
          "username": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 3,
                "pattern": "^[a-zA-Z0-9_.-]+$",
                "description": "The username of the user."
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "password": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 3,
                "description": "The password of the user."
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRoles",
                "description": "The role of the user."
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UserPatchBody"
      },
...
      "UserRoles": {
        "type": "string",
        "enum": [
          "admin",
          "user",
          "guest"
        ],
        "title": "UserRoles"
      },
      "UserSchema": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "The unique identifier of the user."
          },
          "username": {
            "type": "string",
            "maxLength": 128,
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9_.-]+$",
            "title": "Username",
            "description": "The username of the user."
          },
          "role": {
            "$ref": "#/components/schemas/UserRoles",
            "description": "The role of the user."
          },
          "lastLoginAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastloginat"
          },
          "credentialVersion": {
            "type": "integer",
            "title": "Credentialversion"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "role",
          "credentialVersion"
        ],
        "title": "UserSchema"
      }
    },
    "securitySchemes": {
      "OAuth2Token": {
        "type": "http",
        "description": "Bearer authentication with JWT tokens.",
        "scheme": "bearer"
      }
    }
  }
}